Initial work on Markov Chain stuff in rust.
This includes what I had when I did the presentation, plus a conversion to Rust 1.0, plus adding in readline.

Due to the conversion swapping iterating over words for splitting on spaces, the tests don't pass and it is much less functionaly.

Also, one of the tests was non-deterministic to begin with.